+++ src/components/ui/checkbox/Checkbox.module.scss
@@ -48,9 +48,7 @@ button {
     fill: transparent;
   }
 
-  &.checked {
-    border: 2px solid transparent;
-
+  [data-state='checked'] {
     svg {
       display: block;
       fill: var(--color-light-100);
@@ -62,7 +60,7 @@ button {
       background-color: var(--color-dark-100);
     }
 
-    &.checked {
+    [data-state='checked'] {
       background-color: var(--color-dark-900);
     }
   }
@@ -79,7 +77,7 @@ button {
       background-color: var(--color-dark-100);
     }
 
-    &.checked {
+    [data-state='checked'] {
       background-color: var(--color-dark-900);
     }
   }
