+++ src/components/ui/checkbox/index.ts
@@ -1,1 +1,1 @@
-export * from './checkbox';
+export { CheckboxRadix } from './Checkbox';
